requirements.txt
uvicorn
fastapi
requests
python-multipart
openai
python-docx
python-dotenv